Removed adapter infrastructure for simpler code
