--- lib/Fhp/Adapter/AdapterInterface.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Fhp\Adapter;
-
-use Fhp\Message\AbstractMessage;
-
-/**
- * Interface AdapterInterface
- * @package Fhp\Adapter
- */
-interface AdapterInterface
-{
-    /**
-     * @param AbstractMessage $message
-     * @return string
-     */
-    public function send(AbstractMessage $message);
-}
