--- lib/Fhp/Adapter/Exception/AdapterException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Fhp\Adapter\Exception;
-
-/**
- * Class AdapterException
- * @package Fhp\Adapter\Exception
- */
-class AdapterException extends \Exception
-{
-
-}
