+++ lib/Fhp/CurlException.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Fhp\Adapter\Exception;
+namespace Fhp;
 
 /**
  * Class CurlException
- * @package Fhp\Adapter\Exception
+ * @package Fhp
  */
-class CurlException extends AdapterException
+class CurlException extends \Exception
 {
     /**
      * @var mixed
@@ -21,7 +21,7 @@ class CurlException extends AdapterException
      * @param \Exception|null $previous
      * @param mixed $curlInfo
      */
-    public function __construct($message, $code = 0, \Exception $previous = null, $curlInfo)
+    public function __construct($message, $code = 0, \Exception $previous = null, $curlInfo = "")
     {
         parent::__construct($message, $code, $previous);
